#6 - refactor: 로그인 모달 기능 분리
